+++ lib/src/flutter_crisp_chat_desktop.dart
@@ -217,4 +217,7 @@ class DesktopFlutterCrispChat extends FlutterCrispChatPlatform {
 
   @override
   void setOnNotificationTappedCallback(VoidCallback? callback) {}
+
+  @override
+  Future<bool> isVideoCallsSupported() async => true;
 }